pyproject.toml
[project]
name = "garmin-training-coach"
version = "0.1.0"
description = "AI training coach powered by Garmin data, LangGraph, and mem0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain-anthropic>=0.3",
    "langchain-mcp-adapters>=0.1",
    "langchain-core>=0.3",
    "anthropic>=0.40",
    "mem0ai>=0.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "streamlit>=1.40",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "api", "ui"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
